Add schedule to build every day
Why do this?  Well the action we use to get Hugo will always pull the latest, so we can build the site regularly with the latest version of Hugo (and probably the latest version of Go).
This may be dangerous...Stuff may break (thinking of the old tampleting issues around Go 1.18), but mostly I think this should be fine since I've got a fairly good eye on the Hugo releases...

This also has the added benefit that any future posts will regularly get published, which wouldn't happen if we didn't schedule the builds.